Rename to LoiController and move to map directory
